// src/app/model/Request.ts
/**
 * Resource Manager API
 * The API of Resource Manager web application
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface Request {
    /**
     * The identifier of the request, generated automatically by database
     */
    requestID?: number;

    eventName?: string;

    eventDescription?: string;

    dateFrom?: string;

    dateTo?: string;

    handleBefore?: string;

    handleAfter?: string;

    comment?: string;

    resource?: models.Resource;

    requestStatus?: models.RequestStatus;

    user?: models.User;

}